Refactor Score property and add new repository methods
Updated ExamResultConfiguration and ExamResult classes to change Score property from double? to decimal? with precision 10,2. Added new methods GetByUserAndQuestion and GetByIdWithDetailsAsync to AnswerRepository and IAnswerRepository. Changed return type of GetAverageScoreByExamAsync in ExamResultRepository and IExamResultRepository from Task<double> to Task<decimal>. Updated AnswerService and ExamResultService with additional validation checks and to handle new Score property type. Removed out UserId, ExamId, and QuestionId in UpdateAnswerRequestDto. Updated ExamResultAverageScoreResponseDto and ExamResultResponseDto to change Score property type.

+++ Repositories/Entities/ExamResult.cs
@@ -7,7 +7,7 @@
         public User User { get; set; }
         public int ExamId { get; set; }
         public Exam Exam { get; set; }
-        public double? Score { get; set; }
+        public decimal? Score { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? CompletionDate { get; set; }
         public int? Duration { get; set; }
