+++ Repositories/Repositories/IExamResultRepository.cs
@@ -4,7 +4,7 @@ namespace ExamApp.Repositories.Repositories
 {
     public interface IExamResultRepository : IGenericRepository<ExamResult>
     {
-        Task<double> GetAverageScoreByExamAsync(int examId);        //genel başarı durumu
+        Task<decimal> GetAverageScoreByExamAsync(int examId);        //genel başarı durumu
         IQueryable<ExamResult> GetByUserId(int userId);
         IQueryable<ExamResult> GetByUserIdAndExamId(int userId, int examId);
     }
